+++ init.php
@@ -1,6 +1,7 @@
 <?php declare( strict_types=1 );
 
 use FernleafSystems\Wordpress\Plugin\Mandate\Plugin;
+use FernleafSystems\Wordpress\Plugin\Mandate\PluginIdentity;
 
 if ( !defined( 'ABSPATH' ) ) {
 	exit;
@@ -10,7 +11,7 @@ if ( !defined( 'ABSPATH' ) ) {
 	$mandate_autoload = __DIR__.'/vendor/autoload.php';
 	if ( \is_file( $mandate_autoload ) ) {
 		require_once $mandate_autoload;
-		Plugin::boot( __DIR__.'/plugin.php' );
+		Plugin::boot( __DIR__.'/'.PluginIdentity::MAIN_FILE );
 	}
 	else {
 		global $mandate_unsupported_reason;